chore: update .gitignore and refactor test imports
- Added *.scratch.ts to .gitignore for local scratch files.
- Created dist-test/tsconfig.tsbuildinfo for TypeScript build info.
- Refactored import paths in admin-access.e2e-spec.ts and edge-cases.e2e-spec.ts to use new helper module.
- Introduced test/helpers/test-app.helper.ts for creating test applications.
